+++ modulo2/projeto-whatslab/projeto-whatslab/src/components/css.js
@@ -1,107 +1,107 @@
-import styled from 'styled-components';
+// import styled from 'styled-components';
 
-const ContainerMessages = styled.div` //Main screen
-  display: flex;
-  flex-direction: column;
-  min-height: 95vh;
-  width: 515px;
-  margin: 10px auto;
-  justify-content: end;
-  align-items: flex-start;
-  border-radius: 10px;
-  border: 2px solid gray;
-  word-wrap: break-word;
-  background-color: lightgrey;
+// const ContainerMessages = styled.div` //Main screen
+//   display: flex;
+//   flex-direction: column;
+//   min-height: 95vh;
+//   width: 515px;
+//   margin: 10px auto;
+//   justify-content: end;
+//   align-items: flex-start;
+//   border-radius: 10px;
+//   border: 2px solid gray;
+//   word-wrap: break-word;
+//   background-color: lightgrey;
 
-  container{ //Display box-messages to the right
-    display: flex;
-    width: 60%;
-    flex-direction: column;
-    margin: 15px 5px;
-    margin-left: 200px;
-    text-align: end;
-    word-wrap: break-word; 
+//   container{ //Display box-messages to the right
+//     display: flex;
+//     width: 60%;
+//     flex-direction: column;
+//     margin: 15px 5px;
+//     margin-left: 200px;
+//     text-align: end;
+//     word-wrap: break-word; 
     
-    p{
-      margin: 2px 10px;
-      font-size: 17px;
-    }
-  }
+//     p{
+//       margin: 2px 10px;
+//       font-size: 17px;
+//     }
+//   }
 
 
-  div{ //Layout box-messages to the left
-    display: flex;
-    width: 60%;
-    flex-direction: column;
-    margin: 15px 5px;
-    margin-right: 200px;
-    word-wrap: break-word;
+//   div{ //Layout box-messages to the left
+//     display: flex;
+//     width: 60%;
+//     flex-direction: column;
+//     margin: 15px 5px;
+//     margin-right: 200px;
+//     word-wrap: break-word;
       
-    .nameUser{
-      margin: 2px 10px;
-      font-size: 17px;
-      font-weight: bold;
-    }
+//     .nameUser{
+//       margin: 2px 10px;
+//       font-size: 17px;
+//       font-weight: bold;
+//     }
 
-    p{
-      margin: 2px 10px;
-      font-size: 17px;
-      font-weight: lighter ;
-    }
-  }
+//     p{
+//       margin: 2px 10px;
+//       font-size: 17px;
+//       font-weight: lighter ;
+//     }
+//   }
 
-  section{ //Display box-fields: name, message and button
-    border-top: 1px solid gray;
-    display: flex;
-    height: 95%;
-    width: 100%;
-    justify-content: center;
-    align-items: flex-end;
-  }
+//   section{ //Display box-fields: name, message and button
+//     border-top: 1px solid gray;
+//     display: flex;
+//     height: 95%;
+//     width: 100%;
+//     justify-content: center;
+//     align-items: flex-end;
+//   }
     
-  .inputName{
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-    font-size: 15px;
-    width: 6vw;
-    height: 4.5vh;
-    border-radius: 4px;
-    border: 1px solid gray;
-    margin: 10px 5px;
-  }
+//   .inputName{
+//     justify-content: center;
+//     align-items: center;
+//     text-align: center;
+//     font-size: 15px;
+//     width: 6vw;
+//     height: 4.5vh;
+//     border-radius: 4px;
+//     border: 1px solid gray;
+//     margin: 10px 5px;
+//   }
 
-  .inputMessage{
-    justify-content: center;
-    align-items: center;
-    text-align: left;
-    font-size: 15px;
-    width: 100vw;
-    height: 4.5vh;
-    border-radius: 4px;
-    border: 1px solid gray;
-    margin: 10px 5px;
-  }
+//   .inputMessage{
+//     justify-content: center;
+//     align-items: center;
+//     text-align: left;
+//     font-size: 15px;
+//     width: 100vw;
+//     height: 4.5vh;
+//     border-radius: 4px;
+//     border: 1px solid gray;
+//     margin: 10px 5px;
+//   }
 
-  img{
-    display: flex;
-    border-radius: 4px;
-    justify-content: center;
-    align-items: center;
-    width: 30px;
-    height: 30px;
-  }
+//   img{
+//     display: flex;
+//     border-radius: 4px;
+//     justify-content: center;
+//     align-items: center;
+//     width: 30px;
+//     height: 30px;
+//   }
 
-  button{
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-    width: 40px;
-    height: 32px;
-    margin: 10px 5px;
-    border: none;
-    background-color: inherit;
-} 
-`
+//   button{
+//     justify-content: center;
+//     align-items: center;
+//     text-align: center;
+//     width: 40px;
+//     height: 32px;
+//     margin: 10px 5px;
+//     border: none;
+//     background-color: inherit;
+// } 
+// `
 
-export default ContainerMessages
+// export default ContainerMessages
